Code improvements (#30)
* Add customizable scan_interval

The default scan_interval will stay at 3 minutes (180 seconds). In
addition it is possible to define a custom scan_interval in the
sensor entity definition (configuration.yaml).

* Outsource const variables to const.py

- All constants are defined in const.py
- Use CONF_HOST from homeassistant.const

* Use homeassistant.const for device class and units
* Provide icon for particulate matter sensor

Provides icons for PM10 and PM2.5 sensor. Other not so
spread sensor values are not considered.

* Add entity unique_id
Entities can now be configured in Configuration->Entities

* Update readme
* Add “iot_class” to manifest.json
* Bump version to 1.10

Co-authored-by: Joachim Engelmann <joe.engelmann@googlemail.com>